trying to implement the rating component for users to rate games. the view is working (kind of)
implemented the logic so that it should
write new rating to the db. but nothing is getting
written to firebase. also note that only the first
game allows a rating to 'stick' after the user
clicks on stars. this is a pretty good hint
for debugging, though i am uncertain what
it means.