reworked some stuff in the cloudDB. each console maker (nintendo, sony, microsoft, pc) has a field in the document that contains an array, each element in the array is a mapping that contains data about the console. name, full name, release date, type, generation, could have more like description, games, image, but for now this works.
when creating a new game as admin, the consoles
that appear when the console maker is selected
are now links to a new component 'view-console'
similar to view-game, etc.
it also has a tooltip that shows the qualified name
scroll over PSP => PlayStation Portable.

need to debug/ redesign what we did with gameInfoService.
experiencing some loading problems probably
because game info is no longer an observable.
this also results in console being undefined
in view-game component.