switched create-profile to reactive form, ditched the custom form-control component its too much of a pain doing the research all over again for reactive forms.
in the process of changing to a newer technology,
backtracked a little. (havent seen this happen before have we?)
but honestly i feel better and think we will
be out of this dilemna soon enough/ better for it.

have some surface level debugging to do,
then work out saving to db, create v. update.
delete, refrence to user from accountSetting
and reflecting uptodate realtime data in the form when editing.

on the matter of how to refrence user collection.
im thinking it might be best to just create the
account with a blank profile, habing user info
nested within that. but please explore options.
dont just lunge at whatever makes sense in the moment.