got the radio button to work, but eliminated both radio and checkbox custom formcontrol components.
we dont need these. also radio is communicating
data to the form -> db, so eliminated the checkbox
because its really not nessecary / appropriate for this inputanyway

have regressed on the decision of using username
as profiles document id. bad idea.

need to switch that back to uid,, consider re design of db.
so that user data has profile nested within.
that would greatly simplify things

have some routing/ precedence problems.
profile and view-console are unreachable.