Created two UML diagrams, class diagram and sequence diagram for the forum.
Added and modified a bunch of interfaces
so we are pretty much good to go/ in
alignment with the class diagram (withstanding the forum classes)

I may have introduced some breaking changes with this.,, be wary.
Changed a few interfaces with lowercase
file names to titlecase, and didnt
realize vscode has a built in shortcut
for safely doing this until half way thru
so watch out for errors about
incorrect path to interfaces.