GameInfo service is finally doing its job. Made 2 methods for each gameInfo attr: find and get_array. These take the observable and return either the complete array mapped from the firestore doc, or a specific member.
Now that we've got that, the game-index
is non-sturdily working (not super reliably)
but this is the same problem we have with
observables anywhere in the application.
(buffering of external data stalls frequently)

Will need to update a ton of things (i think)
now that the way we access gameInfo
has changed (simplified). Id love it
to find a simple, reliable access schema
for firestore observables. So far they
have really caused a lot of unnessecary
headaches.