Did some quick and easy repairs on preferences and profile edit forms. Changed "create" method to "save" and each of their services. Now its capable of creating new documents for new users as well as updating. Also added a conditional to ngOnInit for each of them, creates a new object that stores the form data implenting their respective interface.
Have a slight roadblock with resolving
pernicious bugs in each of these forms.
We'll need a css loader screen so that
the observable buffers prior to the markup rendering in the DOM.
Secondly, less an issue but still needing resolution-
edit-profile is template driven form
but edit-preferences is reactive (and a use case justifies this)
but for now, I havent been able to figure
how to show previously entered form data
in the form with reactive forms.

Added a stand-in loading screen.
Doesnt cover the content like its supposed to. It looks pretty bad.
But not finished. Its purpose is to allow
observables to buffer in the dom so
we dont get blank images and text fields.
Will need to add animation and nicities,
in addition to covering content completely,
styles.css is colliding with this goal now,
as its supposed to.