Didn't get much done today. Worked out some kinks and simplified dropdown. Made a cooler loading animation, but not 100% where we want it.
added a link from the dropdown to contribs, contrib dashboard
and contribs.

Going to be reteaching myself from scratch, fundementals
of css, sass, less, so that I stop running
into styling roadblocks.  need to get
this site looking professional
and responsive asap. So recent fourthcoming
commits will be limited or
not impactful to the final product.