Got rid of the loading screen for now, it was being glitchy and unpredicatable.
working on a general debug + stars/rating component.

Moved the ability to rate a game inside
the view-game component. Ratings
were not effecting the intended game
otherwise. We need to get an existing
rating to appear when user clicks on a game. Also, we should have
appUser return guest if user is not logged in, and not allow guests
to rate games.