Rating (stars) component is seemingly fully debugged, can now: as signed in user: -rate unrated games -edit game already rated as non signed in user: -unable to submit rating -get routed to sign in page
Other than that, filled in a lot of holes
with routing, routerLinks throughout
site (mainly on dropdown menu)

Everything which im aware of now can
be navigated to by clicking a link
somewhere on the site.

That leaves us with a lot of unimplemented components
and also some thought into redesign
with class diagram. Lots of the site
map has changed and im sure new
classes need to be reflected in our UML