Working on debug of review componentry.
My test account cant view reviews its
created because the username comes
from google, which allows usernames
with spaces in them. our site does not.

So I added display name property to
User interface, in addition to username.
one for human readable one
for machine readable, This should solve
it but we will need to delete and recreate
our existing reviews for the test
account.