Working on a comment section.
Added real time view of ratings and
reviews to contrib-dashboard.

links to reviews are dead links as of now.
Starting to get second thoughts about
using gametitle as part of rating
and review PK- whitespace will cause problems.

need to do some testing regarding this
specifically to find out if we can roll
with the way we have it now (before
making a mess that doesnt work AND
relies on unfounded assumptions).

Its not impossible to make game title
and content id part of the rating or
review interface, but make the PK
a hash value like before. consider this...

minor style changes were made to
display-ratings component.
its now unaligned with heading
in view-ratings. easy fix for next session