Made a TODO file. started generating ideas of how to fix the PK issue with ratings and reviews.
Started working on the comment section.

The skeleton is there but its not
connected a db at this time.

right now the add_new_comment
method is just  a prototype impl.
It pushes newly generated comment
objects to a generic array.

Both the object and array need to derive and contain comment objects
in the finished version.