trying to get the olde gameInfoService working with our neat new component game-info-form.
Since gameInfoService is racked with
boilerplate and tightly coupled with
numerous components in gaming-index
system, created a new service 'testing'
to try to get the game-info-form to play
nicely with an observable of type gameDescriptor
, turning it into an array of objects.

No luck so far. Posted on staxOverflow
to fish for suggestions on a way to best
implement what I was trying to do with
GameInfoService.

tried a different approach by making each doc in GameInfo its own
collection, to me this makes the most sense as documents are to arrays
as collection are to objects/dictionaries,
if we want to take path of least resistance, using a lot of collections
may be the way to go. results:

With the use of the categoriesCollection
approach, we can display the objects
in cloudfirestore db array in our form,
only problem being that these values
are not making it to the formcontrol,
shows up as null in the submitted form,

still this is a good first step, that will
soon require oodles of noodles worth
of work if we decide its the right impl.
consult staxs before moving forward.