Implemented edit functionality for post component and parent component. Changes trace all the way back to db and are reflected in the app. sort of irrelevant to the title of this branch. This is going to require some cleaning up. ie edit gameDescriptor (pending below mentioned design decisions), make method arguments in post, game-info-form and service uniform.
This is going to require some design
rethinking until we arrive at the
reasonable implementation. The
problem is that the game-info
form only implements a single method
for edit, add and delete, yet various
documents possible for writing to,
and individual services for each,
need to think of some unifying logic,
probably at service level, to clean this all up.