Though of a design solution for game info service.
using testing.service to test, all looks good. implemented one use case
in game-info-form, for getting all
documents of a certain type. it works
as expected. thank goodness.

Will require further experimentation
and impl of other use cases
before we determine if this
is the right solution for moving forward