In terms of viewing observables either as an array or a single document, our new impl of game-info is vastly superior- No more massive 5 second load times whenever looking at game-info.
all we need to do now is show that
CRUD can be supported in game-info-form. After doing this, We can build
a real data store for replacing
the massive nested game-info cluster,
as well as moving the functions from
testingService to gameInfoService.

Hoping we can get all this done by tonight and move onto client facing
business and implementing admin and
normal user notification systems.

worked out a bug in preferences. seeing
to it that observables can be preloaded
to the form, dont see a reason why we
cant implement reactive forms sitewide.