Moved TestingService out of effect and into gameInfoService, changed all components that relly on this service so things are up to date.
modified the cloud db so that data
is stored the way that gameInfoService
expects it to be.

updated the game info form so that
each type of game info has a input field
and displays all game info of that type.

Probably dont need to implement
the ability to limit the number of
items in the Posts component list,
since the amount of entrys for
each type of game info I expect to be
low (less than 10, each).