Working on game-info-form and child post, grandchild descriptor-form.
Finished product will allow users
to edit, delete and verify changes by
displaying them in the same field used
for providing input, augmenting the
current iteration with validation and
the ability to provide a description
through a dropdown menu.

Hung up right now on editing. gameInfoService getDocument$ is being
unreliable and not able to identify the source of problem. Examine
sub components of gaming-index
(categories, creators, console_makers)
as these use the same method and
actually fetch the data.