Made a bit of progress on Descriptor form, Post and in effect Game-info-form.
pressing edit now populates the top
form (descriptor form) with the
data for updating. Added some
validation and component logic
and event propogation  so that
these chages can be saved to db.
not quite there yet, the event prop.
chain has not quite reached
game-info-service. wont be long
until we are moving on to bigger
and more important matters, but
alot of important angular features were
learned in making this form.