did some small testing.debug and cleanup
console_makers got changed to
platforms site wide as that seems to
make more sense for the term.

trying to display game info in
the admin home page. not working
so far with unknown cause of this
problem.

Next we should make a form component for adding, editing and deleting
consoles. and the UI sitewide for
interacting with these console choices.

if not that, building a notification system
that compliments the variety of needs
for this site should come next (threads,
comments, content notifications, admin
notifications, suggestions, approve
deny moderators to forums.)