Added C.R.U.D support of the consoles collection in Game-info service. I put it there because the project file system is getting very bloated, and I consider consoles to be a part of game info- something descriptive about a paticular game, not logically important enough to garner its own service, but keeping it as a nested object within game-info would be too complicated.
Layed the groundwork for a more sophisticated version of the post
component seen in previous commits,
without making it a reusable, nested
group of communicating components.

right now reading to the database is
working, using getConsoleMaker$ method for each maker/ platform.
Next, we will make a dropdown form
for adding, editing and deleting
consoles. Note we will probably
need to expand on the number of
fields within the VgConsole interface to
support the information we need the console object to carry.