console-info-form is working almost 100%
got edit button working pretty much
completely (minus some nicities)
after updating the db, the form isnt
reset and closed like we'de like it to.
This is likely because the opening of
the form and population of form data
is triggered by component change detection , likely triggered
when the db values update. well need
to review angular basics to figure out
how to interupt that.

since we arent testing this component anymore, except the maybe
(hopefully) 5 minutes required
to fix the above mentioned problem,
removed the minimized component from the homescreen. we will now need
to navigate to game-info-form and scroll
down to find this.