Seems we have debugged everything with both game-info-form and console info-form. can effectively create, read update and delete. left the <p> element that displays editing mode at top of both types of posts. for now, for debugging purposes. will be using ths interface to do data entry, providing each released console for each type of platform, that will be painstaking.
made some minor tweaks:
-console-form now displays the correct
console image per each entry.
-added validation so forms wont submit
unless required form controls are provided.
-categories, creators and platforms are
in headings instead of labels now.

future tasks- make both these crud systems less hardcoded. if for
example, a platform is added or
deleted, the console-info-form wont
be updated to reflect that change,
because you hardcoded the components
to retrieve data for sony, nintendo, microsoft, pc, web and mobile.

add popup info to both forms so that
the form tells users exactly what is
provided, for now it just highlights an
entry field that is left blank, it gives
no info on how to correct  it.