Working on the forum. Progressing slowly.
started from scratch , needed to rethink
a lot of design decisions.

added some css, then decided we
should probably wait til all the features
are done before worrying about that.

created the shape of the forum homepage: nav bar, thread family
grouping/ body, side bar with metadata
and forum info.

Working on create-thread component
now, realized that we dont need to
collect the family info as with forums you
click on a topic, THEN choose to make
a thread within that topic. the way we
were doing it previously leaves the user
with too many choices.