rebuilding forum. Can't figure out what took us 4 seconds to do with the previous iteration (using template driven forms)
tried trouble shooting the reactive
reusable components, with no luck.

Have a bug with loading thread
observable. Have def seen this before,
with games component,
you follow the link, which works
in the link bearing component
(shown in tool tip when you
hover over the link) then it loads
the link but without the thread.uid

console says thread is undefined, so
tried a bunch of different strategies for
defining it diferentl;y. no luck
come at this tomorrow with a refreshed
mind, hopefully wont stall us for much
longer. lot of work to be done with forum
in paticular, the rest of the site in general.

Closed one issue today, because we
misunderstood the problem..