Did a good chunk of background research on how to develop a forum. Turns out its a task the big game sites outsource to Saas.
Determined that our effort on building
the forum client side is misplaced. We
need to focus on admin first to build
up the forum structure.

Added a service for Forum_Info and
collection in db with metadata,
describing the types of forum meta data-
categories, prefixes and types.
Made an admin forum for editing these,
it uses the same mechanics as game_info
form so little buggy at this time, hoping
to work out some of these kinks.

Forum info form is now working but the
db entries dont get updated when we
update the forum info list. plus the same
old bugs from post component, which
seem to be about editing mode and the
reset and edit buttons on post and descriptor form component.