Knocked out a ton of refactoring, the rest of the refactoring will be step by step ( STEP BY STEP) module by module, component by component, working out all the kinks so that components that need a path routing to them are routed to.
The question of starting anew (without
erasing all this material) and copy and pasting is still on the table.
We have just run into bug after bug
trying to get the site to run
since adding the app.routing
module and feature modules.
These bugs mostly all have self
explanatory causes, and were mitigated
when we made a shared module and
imported it to each module.
However, they are still popping up
aggressivly left and right, mostly
pertaining to routing and forms (which
the site depended heavily on)