Still dealing with a total mess here, but we managed to salvage our dead father from the belly of a whale.
The site is back to previous state- lost
some big changes including work
on form and game-info-form.
dissapointed by this but going
to keep editing until we are more or less
up to speed. its quite overwhelming,
but currently copy and pasting from
the up to date master branch in all
the places where we lost componentry
or services, etc.

let this be a lesson that the next
time you want to completetly update
the architecture of your application,
make a branch PLEASE.