Lot of refactoring, copy and pasting, debugging.
Got the site fairly back on track,
routes are stil a mess, app module
is crammed full of routes,
imports and declarations.

a lot of routes still do not work,
admin and forum in paticular
(what we were working on before
we got into this here mess)

next steps would be merging the current branch to main, making a new
branch and possibly some issues,
so that we can SUCCESSFULY
INCREMENTALLY and with a backup,
make each part of the site its own
feature module complete with functioning routing. Will probably
need to adandon the desire
to make each module
have its own app.component
just do what is practical and nessecary
at this point.

probably willl want to work on auth
module first. that or admin, next forum