Added an AppRoutingModule, where feature modules are lazily loaded. stilll working on admin module, not sure if we are using the up to date syntax for lazy loading, found two.
reinstalling node modules
to ensure things are up to date.

updated admin routing and admin module so that
each component used is exported
and imprted to admin module,
admin routes module treat route prefixed
with 'admin' as root ' ',
and children are included
in children array.

not working but
very close. we got
a lot done today
(giving self a pat
on the back-
considering
project revived,
inspiration to finish
up is through the roof