Properly configured root component, routing, and module imports/exports for -admin -user -content -contrib got started on forum, didnt finish
trying to decide if we want to
keep service main module
imported as root, make
a core module for services,
authguards and models,
or inject them manually
wherever theyre needed.

The last approach seems
to me to be quickest,
but also not efficient.
the current import scheme
for these dependencies is a bit
wonky