Making slow but certainly steady progress towards our goal.
Created a core module, which I
mistook the purpose of. IT only
provides auth service to the app,
similar to the react auth.context:
```
<Auth.Context value={v}>The rest of your app</Auth.Context>
```

Im sure there will be many breaking changes as we were
pretty reckless to get to this point. meaning ->
we moved models and services around from src/app to this
module, thinking the point of
it was to have a common place
to make all shared declarations and providers - alas we already
have a module for both of those!

-got rid of common module since besides services, this was
mostly stuff that goes in shared.

-validators is now moved inside
of services since it provides a service. sub folder
for auth-guards in here as well?

so we need to get rid of many
imports and replace them with
url w/o 'common' as a token.

Other big bug, mostly relating to this is that there are many
constructor arguments which
are invalid because the
import url has changed.