Forum module is our project today,
it was working slighly before the site had the module maelstrom.

Hope to get it back
to that state, but
improved, and with
better ui, using
angular material.

Worked on routing,
this will probably be
the trickiest feature
 module w/ routing
 because of the
 possiblity for nested
 forums. Right now
 not very happy with
 the way its set up,
 so im suggesting
 we look at some
 examples and
 carefully plan out
 the routing
 without much as
 a worry about
 connecting the
 module to the rest of the site (at this point) (ie dropdownComponent)
 stop editing dropdown in every commit, because
 the routes need to
be clean, debugged and SEoptimized
before we can have
anything in concrete.