Set up storage and file upload methods, in the process made site-dashboard even more of a mess than it was initially.
not a good set up we have here- only this
parent component has the content services
as dependencies, so were going to end up
cramming a lot of lines in here unless we
can break down the features MORE.

As of now, we have the dialog returning
the file object, then it gets uploaded, and we
can put the download url in a img src attribtue once it resolves.

tricky little manuever here, why i reccomend
learning more about async js and Async
Await, because  we need to save this
downloadURL in the firestore db when we
save the form return value, however,
the upload to fire storage isnt completed until way after this happens,

anyway, big note- this code probably
belongs in a service duuh,
so consider that when planning how to
make the site-dashboard ts file less
bloated and responsible for every little thing.