Tried to get a lot done today, but didnt really. Made two new components for breaking down code complexity in site-dashboard.
upload-image component will handling
uploading images to storage and getting
downloadURL and progress

mat-data-table will handle wehter to show
a minified or maximized, feature rich mat-table. ITs not working now
because I'm pretty sure that we did too
much development with not enough testing

Have this complete set up for
mat-data-table in place of
recently added streams table,
but its not working, nothing is displayed,
when we choose to show the data-table,
wondering if something got lost in the process of passing in all those
arguments.

Need to solve a async problem to get
upload-images component working,
it needs to return a promise of
uploadPercent AND an observable of
downloadURL, and we need to figure
out how to get those sequentially
emmited values added to an array
for the process of uploading multiple images.