Created a component for uploading multiple images, with that came upload-task.component and dragzone.directive.
Havent yet integrated this with any
of our mat-dialog components, or
gotten it working with single image
upload, but that should be simple.

Right now, trying to get the delete event
working with the upload-task component,
so that when user removes an image from
storage, the results table in the form
is updated. Need to find a directive
or some way to tell the table to re-render
when the array its looping over is updated,
I figured angular would do this for me
automatically but i was wrong.