More functionality of the upload-image and upload-images components working.
Still havent got uploaded images to save
to database, the event hasnt been properly
emitted up the chain.

Also realized we need two things,
1. Automatic delete of all images from
storage when the user doesnt submit the
form during editNewsDialog new news or
edit news dialog.
2. a global service which compiles
list of "in-use" download URLS,
urls that the site needs to display content,
and delete all images in storage which
arent on that list of in use downloadURLS,

TODO; make a git issue about this
so that we dont forget, failure to do
this when the site goes live/ scales
at all would be highly financially irresponsible.