Fixed up upload-image component, so that you can delete a titleCardImage, while editing a preexisting object.
Also added storage cleanup to deleteing
news object from the site-dashboard.

This has a small problem having to do with observables. Posted stax ex
question, to crowd source a solution,
hopefully someone gives a decent
answer because We run into this
problem again and AGAIn-
basically trying to do some procedurally,
after subscribing to observable of doc
in database, results in error,
because the obervable hasnt
returned a value before we
try to do something with the variable
containing the document data,
right now, this problem is in newsService-
& site-dashboard- delete from
storage and firestore works, but
ONLY after you click delete a handful
of times