Still just inching along with this  image upload componentry:
 -had both upload image and images
 working, in functionality and appearance:
 -only uploads were showing on create
 -only previous uploads were showing
 on edit

Then we realized, storage needs to be cleaned ONLY if user submits form,
after deleting an image, should user
submit form, delete the img from
storage. Otherwise, they could delete
the image and not submit the form,
leading to the image url existing in
form and database, but deleted from
storage. Now, the task and delete buttons in
the upload components just emit events,
up until edit-news, which compiles a list
of all image urls which the user chose
to delete. Then when the form submits,
we do the deleting from storage.

Also debugged a problem manifesting with content-dropdown component ,
caused by a small screw up with the
timestamps (we were adding update
timestamp on creation and creation
timestamp on update, meaning
there is no createdAt timestamp when
the news has been created but never edited