So sick of being stuck on debugging this image uploader componentry. There are an endless number of possible problems.
basically got it so that both uploaders are working
 and delete when you press delete button.

 our previous problem flip flopped.
 We took completed update status bar
 out of upload-task and made a single
 one in each upload-image and upload-images.
 now when upload is complete on initial
 upload, we no longer get a confirmation
 that the upload was successful, on upload
-image. BAUMY. but delete works,
both inside and outside the dialog, though,
not every messy possible problem is
accounted for. but clean up is nearly succesful.

Think we can call it quits for this, for now,
move onto other form controls, or another part of the site entirely.
maybe leave this project for after you get a job.
I think having a working prototype of this site would be helpful.