created a sub form control component for multi- select chip list. It works with form control. WE also added a multi text input for dialogs. the remaining two input types- checkbox and radio, we implemented the displays for but were not able to find a quick solution for tying to a form control, we are going to leave that for now, since there is no use case for either of those at this point, that the site needs.
would be interested in comming back, to make these
types work, in addition to fleshing out the multi text input
to optionally support full crud with a nice UI, that way
we can replace the buggy, problematic descriptor form
used throughout the admin module.

Next we should move on to making dialog component
take input and give output, so we can put it to
work, replacing edit news/ streams/ watchlists
in site dashboard. We made a mistake last time
of tying the strategy - uploading images to storage
to the display - dialog form.