DIalog component is more or less reusable. few minor adjustments need being made. We also replaced the og markup and removed testing scripts in app.component/app.module
got dialog component working for watchlists
only in site dashboard, it can successfully
create new objects in watchlists collection.
no functionality for edit or delete at this point.

also havent found a good place for how to
get the data needed for building a dialog
into the config. It needs to be done at some
point, but where is best from a design stand point
is not yet determined. Think long and hard on it.

We also might consider leaving images/storage out of
the dialog entirely for this stage in things, while
it would look good, it is a big time sink and (sort of)
a nice to have feature (for a prototype at least)
maybe spend a commit or two tomorrow trying
to get it working.. certainly not all day.