Made a new component "crud-dropdown" within the site-dropdown directory. It pretty much eliminates all the heavy lifting from site-dashboard, which now mearly handles the setting up the dashboard UI, supplies a list of content options for looping over.
Being stuck on a small, irritating bug which came from
having a small distiniction between the singular and
plural uses of the content types, without a clear
distinction of when to use one or the other, made
me want to learn debugging. this was something
I never learned in school, but realize could
save me hours of manual debugging labor.

I forward this proposition. Also, the annoying
distinction mentioned above was hardly resolved,
just got working enough to move on. I propse
we just cut off the s on watchlist's', stream's', review's', as
well as add the 'official' distiction to 'official reviews', on
components where the type info is displayed to
users of the site. Avoid the headache of naming
mishaps whever possible.

in addition to huge amount of refactoring/ eliminating
redundant code, we have creation of content with the
dialog box working for all types of content.

these now also use a nested object for metadata, which
stores info about the creator(S), when it was created
and updated, the type, etc.

We need to make the ability for the crud component
to detect contributors more robust in two ways
1) it gets the signed in user from localhost, when
we could use auth service instead.
2) if multiple people edit a news article, it should
be able to tell, changing the sole contributor to an
array of usernames, rather than overwriting the existing
like it currently does.

with the addition of metadata to the content object,
we are able to more reliably sort the various types of content.

We also enabled deletion of content with the crud-dropdown.

after adding ability to update content, as well as the
nuances mentioned above, we can move on to other
parts of the site, or remain here to flesh out this
component in greater detail, seeing it will be reused
quite a bit in my estimations