Had an idea for the metadata part of content crud. important idea, that we dont really need to involve the dialog component with metadata at all.
so in the process of implementing this-
a process for updating metadata,
I realized i cant acesss user data from auth state
outside of the core module (which dropdown is in,
the only component that actually fetches
that data reliably). Really fed up at this point,
because I (somewhat naively) assumed we were
done debugging authentication and the nuances
of lazily loaded feature modules.

need to do some more reading today. determine
if there is something I'm missing about this modular
architecture and how the core module works.
I was really hoping, when I first learned about
core module, that I could set up my auth state and
have access to a user object wherever I go in the site.

I want to look into NGRX and state management,
its something that is going to be needed for an
app of this size.

two todos:
-research core module and how to globalize auth state data
-ngrx, learn it, is it right for this project? how much
backtracking would we need to do?

feel conflicted on the topic of backtracking. I have invested
loads of time and effort into this project. but in retrospect,
its not much the current final product. I dont want to
waste that effort, but also dont think it would be a terrible
idea to wope the slate clean: meaning just keep the base
routing and modural structure, maybe dropdown and css.

there is just so many bugs and you havent been doing
any testing or serious debugging. This really isnt a good
exemplar of quality software engineering. You need to find
a way to take this project to the next level if its going to
worth continued effort in the long run. and rememeber
just because you wipe the slate doesnt mean you cant
pick and choose the good working components from
the older versions. just something to think about...