not even really sure what is going on at this point. There are two repos in the workspace with the same name, when we tried to clone. one is the branch and one is master.
I had planned to delete the contents of the branch
and copy and paste the new working project inside,
then merge, but it seems git is doing some overhead
behind the scenes. going to commmit this and see what
happens. I am theorizing there are two copies exisiting
concurrently, both on master and branch.