Actually things came together quite nicely. we had to delete node modules, clear npm cache, reinstall, because of a checksum integrity problem (have an idea of how that came about...) but all things considered, merging the old and new repos by hand could have been a lot more troublesome and fraught with peril.
Now onto design and development.
Please don't sleep on the design part.
have a plan for all layouts, the overall aesethetic,
the custom aesethtic of a feature module, and also
really understand the strengths and potential shortcomings
of ngrx. Dont just jump in assuming everything will
be great and work out in time. thats how we got
in the just overcome mess in the first place.