Laying the ground work for transfer of our prototype into the working app. We created an index file for effects and one paticularly in admin module, to export all the effects that admin module needs.
We will need to set up a game-info and forum-info
store chain (not sure what its called: actions + reducer + effects).
because the creation of new games or forums requires
the user make a selection to determine the format
of the forum. alt, we could simplify the complexity
of both of these forums, like we did with content.