Added ```UserActivityClassDiagram```, which is pretty much a new version of the site class diagram. realized we have a lot more planning/ design and layout prototyping to do before we continue implementing
then continued implementing (facepalm)
knocked out some actions for smaller details
of the site, then realized the way we have this data
organized in the old site is way unorthodox.
remembered the rationalization behind that,
and that this probably needs more consideration-
because game and forum info isnt
stuff that is modified super often, or by users,
we keep all types of game info in a single collection
and sort them at component level.

as such, the individual attributes for our interface
are merely sub collections for types of game info.
So i'm not totally convinced it makes sense or is ideal
to use ngrx for this. ideally the whole site should be
covered by redux design pattern (if practical) but
this does feel strange and I had to stop. before
making a big(ger) mess. NEEDS FURTHER CONSIDERATION!
***add it to the list***