Thought we could get game info working if we rethought the way data is stored in the document
basically we are treating it like a collection (as intended)
by removing VG_console typed objects, and only storing
gameDescriptors, now renamed GameInfo.

We are really stumped on why the query for gameInfo
wont work. and VERY tired of being stuck on ngrx
and rxjs observable bugs. I think its time we take
a break from this project. but you also need to
learn debugging in general. youve been saying this
for months, and then proceeding to do nothing.

not ready to give up, but i am quickly losing patience.
maybe it is soon time to return to python land?