Project is currently kinda broken, but fortunately we found the source of the problem.
That was because we were using an index.ts file to export
reducers and effects, and import them into the feature level
app.module, which was not working. thus content entities
were populated, because that store was set up in content
module alone. The problem is setting up two sources of
truth / stores in the app module. Not sure how that
causes a firebase bug. We also commented out the
other effects besides query and that got rid of the firebase
error, but to avoid forgetting we did that, un commented it
just now. If there is bugs with GameInfo after this refactor,
you know where to look.

project file/ folder storage has changed a bit , causing
the anticipated breakage. We now have a centeral
store folder, which contains folders for effects, actions
reducers, models (in near the future). along with
setting up the index.ts files in each folder and sub folder,
we will also need to adjust the import statements in
each feature app.module, ensuring that features get
exactly what reducers and effects they need.